【WA】[AtCoder Beginner Contest 030]D - へんてこ辞書

+++ abc030/abc030_d.cpp
@@ -44,7 +44,7 @@ typedef priority_queue<ll, vector<ll>, greater<ll> > PQ_ASK;
 
 
 int main() {
-    int n, a;
+    ll n, a;
     string k;
     cin >> n >> a >> k;
     a--;
